llm_eval_pipeline/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
click>=8.0.0
aiohttp>=3.8.0
tqdm>=4.65.0

# LLM API clients
openai>=1.0.0
anthropic>=0.15.0
groq>=0.4.0

# ML/AI frameworks
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0

# Evaluation frameworks
inspect-ai>=0.3.0

# Visualization and reporting
plotly>=5.15.0
kaleido>=0.2.1

# Testing and development
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional dependencies for enhanced functionality
# Uncomment as needed:

# For advanced NLP metrics
# nltk>=3.8.0
# rouge-score>=0.1.2
# bert-score>=0.3.13

# For database storage
# sqlalchemy>=2.0.0
# sqlite3  # Built-in with Python

# For distributed evaluation
# ray>=2.0.0

# For monitoring and logging
# wandb>=0.15.0
# tensorboard>=2.13.0

# For advanced visualization
# matplotlib>=3.7.0
# seaborn>=0.12.0

# For statistical analysis
# scipy>=1.10.0
# scikit-learn>=1.3.0

# For configuration validation
# pydantic>=2.0.0
# jsonschema>=4.17.0